+++ ecmascript/transforms/proposal/tests/decorators.rs
@@ -1,9 +1,11 @@
 use swc_common::chain;
+use swc_common::Mark;
 use swc_ecma_parser::{EsConfig, Syntax, TsConfig};
 use swc_ecma_transforms_base::resolver::resolver;
 use swc_ecma_transforms_compat::es2015::classes;
 use swc_ecma_transforms_compat::es2015::function_name;
 use swc_ecma_transforms_compat::es2020::class_properties;
+use swc_ecma_transforms_module::common_js;
 use swc_ecma_transforms_proposal::decorators;
 use swc_ecma_transforms_proposal::decorators::Config;
 use swc_ecma_transforms_testing::test;
@@ -5496,4 +5498,95 @@ await: function _await() {}
 };
 
 "#
+);
+
+fn issue_395_syntax() -> ::swc_ecma_parser::Syntax {
+    ::swc_ecma_parser::Syntax::Es(::swc_ecma_parser::EsConfig {
+        decorators: true,
+        ..Default::default()
+    })
+}
+
+test!(
+    issue_395_syntax(),
+    |_| chain!(
+        decorators(Default::default()),
+        common_js::common_js(
+            Mark::fresh(Mark::root()),
+            common_js::Config {
+                strict: false,
+                strict_mode: true,
+                no_interop: true,
+                ..Default::default()
+            }
+        ),
+    ),
+    issue_395_1,
+    "
+import Test from './moduleA.js'
+
+@Test('0.0.1')
+class Demo {
+ constructor() {
+    this.author = 'alan'
+ }
+}
+",
+    "
+'use strict';
+var _moduleAJs = require('./moduleA.js');
+let Demo = _decorate([_moduleAJs.default('0.0.1')], function(_initialize) {
+  class Demo{
+      constructor(){
+          _initialize(this);
+          this.author = 'alan';
+      }
+  }
+  return {
+      F: Demo,
+      d: []
+  };
+});
+
+"
+);
+
+test!(
+    issue_395_syntax(),
+    |_| chain!(
+        decorators(Default::default()),
+        common_js::common_js(
+            Mark::fresh(Mark::root()),
+            common_js::Config {
+                strict: false,
+                strict_mode: true,
+                no_interop: true,
+                ..Default::default()
+            }
+        ),
+    ),
+    issue_395_2,
+    "
+const Test = (version) => {
+return (target) => {
+  target.version = version
+}
+}
+
+export default Test
+",
+    "
+'use strict';
+Object.defineProperty(exports, '__esModule', {
+  value: true
+});
+exports.default = void 0;
+const Test = (version)=>{
+  return (target)=>{
+      target.version = version;
+  };
+};
+var _default = Test;
+exports.default = _default;
+"
 );
